4.2.1r11 - Channellist  - fixed picon display  - added new Option "enable second line info without picon display" - added missing recordings to Database, fixed movielist performance and missing recordings in DLNA Server - EPG:  - fixed handling for events with long durations (24h)  - send load_finished/save_finished cachestates after database load/save has finished - fixed parental control for recordings - fixed tuner slot descriptions for tuners configured to loopthrough - fixed CEC handling for "Routing Information" Messages (fixes unwanted "power down" for some cases), add OPPO to CEC Vendor list - fixed "eTimer from wrong thread" crashes - removed resolution check when checken for color LED (should fix external lcds that override /proc/stb/lcd values - log-warn when someone uses doClose directly (you're not supposed to do that unless you really know what you're doing)

+++ usr/include/enigma2/lib/service/event.h
@@ -25,7 +25,7 @@ struct eventData
 		return iso_639_language_code < e.iso_639_language_code;
 	}
 	time_t begin_time;
-	uint16_t duration;
+	uint32_t duration;
 	uint16_t event_id;
 	std::string iso_639_language_code;
 	std::string title;
